A2A/requirements.txt
asyncua>=1.0.0
psycopg2
langchain-openai
langchain-anthropic
python-dotenv
pydantic
langchain-community 
starlette
fastapi
click
uvicorn
httpx
asyncclick